refactor: Remove system namespace from system part
