+++ Assets/2.Scripts/System/DummyMovement.cs
@@ -2,32 +2,29 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace System
+public class DummyMovement : MonoBehaviour
 {
-    public class DummyMovement : MonoBehaviour
+    private Rigidbody2D _body;
+
+    private float _horizontal;
+    private float _vertical;
+
+    [SerializeField]
+    private float _runSpeed = 20.0f;
+
+    void Start()
     {
-        private Rigidbody2D _body;
+        _body = GetComponent<Rigidbody2D>();
+    }
 
-        private float _horizontal;
-        private float _vertical;
+    void Update()
+    {
+        _horizontal = Input.GetAxisRaw("Horizontal");
+        _vertical = Input.GetAxisRaw("Vertical");
+    }
 
-        [SerializeField]
-        private float _runSpeed = 20.0f;
-
-        void Start()
-        {
-            _body = GetComponent<Rigidbody2D>();
-        }
-
-        void Update()
-        {
-            _horizontal = Input.GetAxisRaw("Horizontal");
-            _vertical = Input.GetAxisRaw("Vertical");
-        }
-
-        private void FixedUpdate()
-        {
-            _body.velocity = new Vector2(_horizontal * _runSpeed, _vertical * _runSpeed);
-        }
+    private void FixedUpdate()
+    {
+        _body.velocity = new Vector2(_horizontal * _runSpeed, _vertical * _runSpeed);
     }
 }
