+++ Assets/2.Scripts/System/MainDataManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0846d5bfc6b52482781749d3b6ba0a8a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
