+++ Assets/2.Scripts/System/MainUIManager.cs
@@ -1,0 +1,12 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine.UI;
+using UnityEngine;
+
+public class MainUIManager : MonoBehaviour
+{
+    [SerializeField]
+    private Text _killedEnemyText;
+    [SerializeField]
+    private Text _survivedTimeText;
+}
