+++ Assets/2.Scripts/System/MainUIManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d1d0cf9a3ae6844558f0be0ff5cd2ae6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
