+++ Assets/2.Scripts/System/SmoothTargetFollowing.cs
@@ -1,22 +1,19 @@
 ﻿using UnityEngine;
 
-namespace System
+public class SmoothTargetFollowing : MonoBehaviour
 {
-    public class SmoothTargetFollowing : MonoBehaviour
+    [SerializeField]
+    private Transform _target;
+
+    [SerializeField]
+    private float _smoothSpeed = 0.125f;
+    [SerializeField]
+    private Vector3 _offeset;
+
+    private void FixedUpdate()
     {
-        [SerializeField]
-        private Transform _target;
-
-        [SerializeField]
-        private float _smoothSpeed = 0.125f;
-        [SerializeField]
-        private Vector3 _offeset;
-
-        private void FixedUpdate()
-        {
-            Vector3 targetPosition = _target.position + _offeset;
-            Vector3 smoothedPosition = Vector3.Lerp(transform.position, targetPosition, _smoothSpeed);
-            transform.position = smoothedPosition;
-        }
+        Vector3 targetPosition = _target.position + _offeset;
+        Vector3 smoothedPosition = Vector3.Lerp(transform.position, targetPosition, _smoothSpeed);
+        transform.position = smoothedPosition;
     }
 }
